+++ packages/common/src/index.ts
@@ -168,6 +168,7 @@ export * from './authorization/types';
 export * from './compiler/exploreCompiler';
 export * from './compiler/filtersCompiler';
 export * from './compiler/translator';
+export { default as DbtSchemaEditor } from './dbt/DbtSchemaEditor/DbtSchemaEditor';
 export * from './dbt/validation';
 export * from './ee/index';
 export * from './pivotTable/pivotQueryResults';
@@ -245,6 +246,7 @@ export * from './types/userAttributes';
 export * from './types/userWarehouseCredentials';
 export * from './types/validation';
 export * from './types/warehouse';
+export * from './types/yamlSchema';
 export * from './utils/accessors';
 export * from './utils/additionalMetrics';
 export * from './utils/api';
@@ -252,7 +254,8 @@ export { default as assertUnreachable } from './utils/assertUnreachable';
 export * from './utils/catalogMetricsTree';
 export * from './utils/charts';
 export * from './utils/conditionalFormatting';
-export * from './utils/convertToDbt';
+export * from './utils/convertCustomDimensionsToYaml';
+export * from './utils/convertCustomMetricsToYaml';
 export * from './utils/dashboard';
 export * from './utils/dbt';
 export * from './utils/email';
@@ -817,8 +820,8 @@ export type HealthState = {
         version?: string;
     };
     rudder: {
-        writeKey: string;
-        dataPlaneUrl: string;
+        writeKey: string | undefined;
+        dataPlaneUrl: string | undefined;
     };
     sentry: Pick<
         SentryConfig,
@@ -1176,24 +1179,6 @@ export function formatRows(
             };
         }
 
-        return resultRow;
-    });
-}
-
-export function rowsWithoutFormatting(rows: { [col: string]: AnyType }[]) {
-    return rows.map((row) => {
-        const resultRow: ResultRow = {};
-        const columnNames = Object.keys(row || {});
-
-        for (const columnName of columnNames) {
-            const value = row[columnName];
-            resultRow[columnName] = {
-                value: {
-                    raw: value,
-                    formatted: null,
-                },
-            };
-        }
         return resultRow;
     });
 }