+++ packages/e2e/cypress/e2e/app/pivotTables.cy.ts
@@ -19,21 +19,26 @@ describe('Pivot Tables', () => {
         cy.contains('2018-04-02');
         cy.contains('$236.21');
     });
-    it.skip('Can create a pivot table chart on explore', () => {
+    it('Can create a pivot table chart on explore', () => {
         // Navigate to the explore page
         cy.visit(
             `/projects/${SEED_PROJECT.project_uuid}/tables/orders?create_saved_chart_version=%7B%22tableName%22%3A%22orders%22%2C%22metricQuery%22%3A%7B%22exploreName%22%3A%22%22%2C%22dimensions%22%3A%5B%22orders_order_date_week%22%2C%22orders_status%22%2C%22orders_is_completed%22%5D%2C%22metrics%22%3A%5B%22orders_total_order_amount%22%5D%2C%22filters%22%3A%7B%7D%2C%22sorts%22%3A%5B%7B%22fieldId%22%3A%22orders_order_date_week%22%2C%22descending%22%3Atrue%7D%5D%2C%22limit%22%3A500%2C%22tableCalculations%22%3A%5B%5D%2C%22additionalMetrics%22%3A%5B%5D%2C%22metricOverrides%22%3A%7B%7D%7D%2C%22tableConfig%22%3A%7B%22columnOrder%22%3A%5B%22orders_order_date_week%22%2C%22orders_status%22%2C%22orders_is_completed%22%2C%22orders_total_order_amount%22%5D%7D%2C%22chartConfig%22%3A%7B%22type%22%3A%22table%22%2C%22config%22%3A%7B%22showColumnCalculation%22%3Afalse%2C%22showRowCalculation%22%3Afalse%2C%22showTableNames%22%3Atrue%2C%22showResultsTotal%22%3Afalse%2C%22showSubtotals%22%3Afalse%2C%22columns%22%3A%7B%7D%2C%22hideRowNumbers%22%3Afalse%2C%22conditionalFormattings%22%3A%5B%5D%2C%22metricsAsRows%22%3Afalse%7D%7D%7D`,
         );
 
+        cy.contains('Tables').should('be.visible'); // Ensure the sidebar has loaded before clicking configure below
         cy.contains('Configure').click();
         cy.contains('Drag dimensions into this area to pivot your table');
 
         const dragSelector =
-            '[role="tabpanel"] [data-rfd-draggable-id="orders_is_completed"]';
+            '[role="tabpanel"] [data-rfd-drag-handle-draggable-id="orders_is_completed"]';
         const dropSelector = '[data-rfd-droppable-id="COLUMNS"]';
 
-        // Drag and drop is not working as expected,
-        cy.drag(dragSelector, dropSelector);
+        cy.dragAndDrop(dragSelector, dropSelector);
+
+        cy.get('[data-testid="visualization"]').as('chartArea'); // Using an alias aviod querying the DOM for the same element multiple times
+
+        cy.get('@chartArea').findByText('Loading chart').should('not.exist');
+        cy.get('@chartArea').contains('Is completed'); // Check that the chart updated successfully with the pivot table(containing 'is completed' column)
     });
 
     it('I can save a pivot table chart and add it to a dashboard', () => {