+++ packages/frontend/src/hooks/useColumns.tsx
@@ -1,7 +1,6 @@
 import {
     formatItemValue,
     friendlyName,
-    getFormattedWithFallback,
     getItemMap,
     isAdditionalMetric,
     isCustomDimension,
@@ -11,7 +10,6 @@ import {
     itemsInMetricQuery,
     type AdditionalMetric,
     type CustomDimension,
-    type Dimension,
     type Field,
     type ItemsMap,
     type RawResultRow,
@@ -49,15 +47,7 @@ export const getItemBgColor = (
     }
 };
 
-export const formatCellContent = (
-    data?: { value: ResultValue },
-    item?:
-        | Field
-        | Dimension
-        | AdditionalMetric
-        | TableCalculation
-        | CustomDimension,
-) => {
+export const formatCellContent = (data?: { value: ResultValue }) => {
     if (!data) return '-';
 
     const { value } = data;
@@ -71,15 +61,10 @@ export const formatCellContent = (
                       {index < array.length - 1 && <br />}
                   </Fragment>
               ))
-            : getFormattedWithFallback(value);
+            : value?.formatted ?? value?.raw;
     }
 
-    if (value?.formatted === null && item) {
-        // Null formatting means the formatting was skipped by the backend
-        // so we need to handle formatting in the frontend based on the raw value
-        return formatItemValue(item, value?.raw);
-    }
-    return getFormattedWithFallback(value);
+    return value?.formatted ?? value?.raw;
 };
 
 export const getFormattedValueCell = (